Fixed bug: ownership of files modified by qbzr
If a file was modified or created by qbzr, the ownership of the file was modified into root. Now the users log in the container as themselves, so that all modifications have the correct ownership. This also simplify the connection to the X server: there is no need to temporarily enable local connection since the container shares the same user as the host. Configurations of Bazaar are written in the configuration directory $HOME/.config/breezy and $HOME/.bzr.log, which are now mounted inside the container.